requirements.txt
pytest
selenium
webdriver-managers